move verified domain query tests
