+++ api-js/src/queries/verified-domains/__tests__/find-verified-domains.test.js
@@ -3,16 +3,16 @@ const { graphql, GraphQLSchema } = require('graphql')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
-const { cleanseInput } = require('../validators')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { createQuerySchema } = require('../..')
+const { createMutationSchema } = require('../../../mutations')
+const { cleanseInput } = require('../../../validators')
 const {
   verifiedOrgLoaderConnectionsByDomainId,
   verifiedDomainLoaderConnections,
-} = require('../loaders')
+} = require('../../../loaders')
 const { DB_PASS: rootPass, DB_URL: url } = process.env
 
 describe('given findVerifiedDomains query', () => {
